Coins now deactivate and reactivate
Coins will now temporarily deactivate (both visually and functionally) for as long as their effect duration and then reactivate after that time has passed to be collected again. This is useful for power-up coins so they don't ever disappear permanently which is important if they are required to complete a part of the level.